requirements.txt
# ── Core ─────────────────────────────────────────────────────
python-binance>=1.0.19
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# ── ML / AI ──────────────────────────────────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0

# ── Optional ML frameworks (uncomment as needed) ────────────
# tensorflow>=2.15.0
# torch>=2.1.0
# xgboost>=2.0.0
# lightgbm>=4.1.0
